TABCE/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy==1.26.2
opencv-python-headless==4.8.1.78
Pillow==10.1.0
pytest==7.4.3
httpx==0.25.2
mediapipe==0.10.8
rembg==2.0.50
scikit-learn==1.3.2
pandas==2.1.3
onnxruntime==1.16.3
